requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-filter==23.1
django-internationalflavor==0.4.3
djangorestframework==3.14.0
importlib-metadata==6.1.0
Markdown==3.4.3
pycodestyle==2.10.0
python-stdnum==1.18
pytz==2023.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
zipp==3.15.0
